\name{WLogit-package}
\alias{WLogit-package}
\alias{WLogit}
\docType{package}
\title{
\packageTitle{WLogit}
}
\description{
\packageDescription{WLogit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WLogit}
\packageIndices{WLogit}
This package consists of functions: "WhiteningLogit", "CalculPx", "CalculWeight", "Refit_glm", "top", "top_thresh", "WorkingResp", and "Thresholding".
For further information on how to use these functions, 
we refer the reader to the vignette of the package.
}
\author{
\packageAuthor{WLogit}

Maintainer: \packageMaintainer{WLogit}
}
\references{
W. Zhu, C. Levy-Leduc, N. Ternes. "Variable selection in high-dimensional logistic regression models using a whitening approach". (2022)
}

