\name{X}
\alias{X}
\docType{data}
\title{
Example of a design matrix of a logistic model}
\description{
It contains an example of a design matrix of a logistic model.
}
\usage{data("X")}
\format{
  The format is:
 num [1:100, 1:500] -1.576 -0.476 -0.237 -0.398 0.284 ...
}
\examples{
data(X)
}
\keyword{datasets}