\name{y}
\alias{y}
\docType{data}
\title{
Example of a binary response variable of a logistic model.
}
\description{
It contains an example of a binary response variable of a logistic model.
}
\usage{data("y")}
\format{
  The format is:
 int [1:100] 0 1 0 1 1 0 0 0 1 1 ...
}
\examples{
data(y)
}
\keyword{datasets}
